+++ packages/compiler/test/RuntimeSliceLayout.test.ts
@@ -11,7 +11,7 @@ import * as Type from '../src/Type.js'
 const ascii = (value: string): Uint8Array =>
   Uint8Array.from(value, (character) => character.charCodeAt(0))
 
-const source = `import silk.usize as usize
+const source = `import silk.usize
 struct Token { kind: i32 flag: bool }
 struct Empty {}
 fn scan(values: &[i32]) -> i32 { return usize.toI32(values.length) }
